requirements.txt
SQLAlchemy~=1.4.47
python-dotenv~=1.0.0
aiohttp~=3.8.4
bcrypt~=4.0.1
asyncpg~=0.27.0
SQLAlchemy-Utils~=0.40.0
